misc/HtmlTestSuite/results/test-0434.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Computed value of orthogonal writing-mode on table cell</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="https://www.w3.org/TR/css-tables-3" rel="help"/>
<link href="https://crbug.com/1307976" rel="help"/>
<style>
  td {
    writing-mode: vertical-lr;
  }
</style>

</head><body><table>
  <tbody><tr>
    <td>Test</td>
  </tr>
</tbody></table>

<script>
  test(() =&gt; {
    let td = document.querySelector('td');
    assert_equals(getComputedStyle(td).writingMode, 'vertical-lr');
  }, 'Computed value of orthogonal writing-mode on table cell');
</script>
</body></html>
